Generate release_date property in netkan
