+++ Spec.md
@@ -337,7 +337,7 @@ A typical install directive only has `file` and `install_to` sections:
 ##### comment
 
 A comment field, if included, is ignored. It is not displayed to users,
-nor used by programs. It's primary use is to convey information to humans
+nor used by programs. Its primary use is to convey information to humans
 examining the CKAN file manually
 
 ##### author
@@ -440,11 +440,52 @@ A list of strings indicating the locales for which this module includes localiza
     ]
 ```
 
+##### download_size
+
+If supplied, `download_size` is the number of bytes to expect when
+downloading from the `download` URL. It is recommended that this field is
+only generated by automated tools (where it is encouraged),
+and not filled in by hand.
+
+##### download_hash
+
+If supplied, `download_hash` is an object of hash digests. Currently
+SHA1 and SHA256 calculated hashes of the resulting file downloaded.
+It is recommended that this field is only generated by automated
+tools (where it is encouraged), and not filled in by hand.
+
+```json
+    "download_hash": {
+        "sha1": "1F4B3F21A77D4A302E3417A7C7A24A0B63740FC5",
+        "sha256": "E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855"
+    }
+```
+
+##### download_content_type
+
+If supplied, `download_content_type` is the content type of the file
+downloaded from the `download` URL. It is recommended that this field is
+only generated by automated tools (where it is encouraged),
+and not filled in by hand.
+
+```json
+    "download_content_type": "application/zip"
+```
+
+##### release_date
+
+If supplied, `release_date` is the timestamp when the described version
+of the mod was released. The format is ISO 8601.
+
+```json
+    "release_date": "2019-11-30T14:54:45.995Z",
+```
+
 ### Relationships
 
 Relationships are optional fields which describe this mod's relationship
 to other mods. They can be used to ensure that a mod is installed with
-one of its graphics packs, or two mods which conflicting functionality
+one of its graphics packs, or two mods with conflicting functionality
 are not installed at the same time.
 
 At its most basic, this is an array of objects, each being a name
@@ -625,38 +666,6 @@ functionality. This allows relationships that ensure only one set
 of assets are installed. (Eg: `CustomBiomesRSS` and `CustomBiomesKerbal`
 both provide and conflict with `CustomBiomesData`, ensuring that both
 cannot be installed at the same time.)
-
-##### download_size
-
-If supplied, `download_size` is the number of bytes to expect when
-downloading from the `download` URL. It is recommended that this field is
-only generated by automated tools (where it is encouraged),
-and not filled in by hand.
-
-##### download_hash
-
-If supplied, `download_hash` is an object of hash digests. Currently
-SHA1 and SHA256 calculated hashes of the resulting file downloaded.
-It is recommended that this field is only generated by automated
-tools (where it is encouraged), and not filled in by hand.
-
-```json
-    "download_hash": {
-        "sha1": "1F4B3F21A77D4A302E3417A7C7A24A0B63740FC5",
-        "sha256": "E3B0C44298FC1C149AFBF4C8996FB92427AE41E4649B934CA495991B7852B855"
-    }
-```
-
-##### download_content_type
-
-If supplied, `download_content_type` is the content type of the file
-downloaded from the `download` URL. It is recommended that this field is
-only generated by automated tools (where it is encouraged),
-and not filled in by hand.
-
-```json
-    "download_content_type": "application/zip"
-```
 
 #### Extensions
